requirements.txt
torch @ https://download.pytorch.org/whl/cpu/torch-1.10.0%2Bcpu-cp37-cp37m-linux_x86_64.whl
transformers
gunicorn
numpy
flask
Pillow
